Messed Around with extending this to 3d
Was able to make predictions in 3d quite easily, making visualizations
of the model output in 3d+ is another thing entirely.

Due to the fact that a contour plot in 2d is representing 3d, and doing
the same in 3d would necessitate being able to make a viz in 4d.